Fix picard heapspace error in align rule
In the Singularity image for picard, the default heapsize is too small,
leading to crashes in the pipeline when the rule 'align' runs out of
heap space. The solution is to explicitly specify the heap space that
picard can use, thereby overriding the default value that is set in the
Singularity image.

This issue was reported on capture-lumc-wrapper, see
https://git.lumc.nl/klinische-genetica/capture-lumc/capture-lumc-wrapper/issues/26
for details.