Reorder Snakefile to make sure the tests work
There were some failed tests due to changes in the order of checks in
the Snakefile, which caused the wrong error to be thrown when there were
multiple missing files. This commit fixes that issue.

The KNOWN_SITES argument has been made mandatory for now, and this
requirement has also been added to the test cases.