Replace biocontainer wc with pywc.py
The wc which is present in the biocontainers has a maximum integer size
of just 2^32, which means the counter resets every 4 billion bases. This
commit adds a custom python script to solve issue #39.

While the python tool is slower than wc, it has the advantage that we
only have to parse the bam file once, since it can count both the lines
and the number of bases in one go. That should save a bit on the time.

This commit also updates samtools from version 1.6 to 1.7.