Add threads to fastqc
Hutspot sometimes fails on the fastqc analysis due to Java heap space
memory errors. It is not possible to set the memory usage by fastqc
directly, since this is done by the perl wrapper script. However, the
amount of memory that is set by the wrapper scales with the number of
threads that fastqc uses. Therefore, the number of threads was
increased, effectively increasing the memory fastqc can use.