Make the scatter output dynamic
dynamic is a special keyword in Snakemake that can be used to mark
outputs when the number of output files is not known before execution.
This is the case when using scatterregions, since the number of outputs
depends on the size of the reference genome and the size of the
scattered chunks.

See
https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#dynamic-files
for details