Call GATK baserecalibrate on readgroup bam files
Base recalibration takes a long time to run. By running the
base recalibration on the separate per-readgroup bam files, instead of on
the output of the markduplicates step, we can run these tasks earlier in
the pipeline, and in parallel with the markduplicates step. This reduces
the total runtime of the pipeline.

This commit also adds a test to make sure that statistics for both read
groups are present in the base recalibration output file.