Consolidate collectstats into a single rule
The implementation is a bit hacky, since snakemake does not allow for
optional input files. As a workaround, "." is passed when the bedfile is
not defined, and the collect_stats.py script has been made aware of the
special meaning of "."

Additionally, Click has been removed as a dependency for collect stats,
and the structure of the stats.json file has been updated to only allow
for a single entry of coverage stats instead of a list. This has been
done to match an earlier change in Hutspot where support for multiple
bed files has been dropped.