Add a singularity prefix to the snakemake profile
Multiple (pytest) processes trying to write the same image to
/tmp/singularity can lead to corruption, leading to intermittent
failures in the gitlab-ci tests. By specifying a singularity prefix in
the snakemake profile, the same images can be re-used, so we only have
to worry about concurrent processes writing the same image when a new
image is added to the pipeline.