Rename bedfile to targetsfile
Hutspot now supports two bed files to calculate coverage. One is the
`targetsfile`, which was called `bedfile` before, and holds the targets
of the capture kit. The other one is `baitsfile`, which holds the bait
locations of the capture kit.

It is possible to specify only the `targetsfile`, but if you specify the
`baitsfile`, the `targetsfile` must be specified as well, since both are
required by picard HsMetrics.

Also added a test for invalid configuration files, and shortened the
jsonschema validation error to only show the human readable message.