Simplify the structure of the coverage data
Previously, Hutspot supported multiple bed files to calculate coverage
against. Because of this, the stats.json file had a nested structured
where the coverage based on each bed file was stored, including the name
of the bed file and the gender, according to that bed file.

Since the current version of Hutspot only supports a single bed file,
this structure has been simplified. All coverage statistics are now
directly under 'coverage' for each sample, and the 'gender' has been
moved out of the 'coverage' statistics to the sample level.