Add picard DuplicationMetrics to stats.json
Unfortunately, this adds a limitation on the sample names that can be
used with Hutspot, since the naming of the samples in the multiQC parsed
output of picard MarkDuplicates is partly ambiguous.

This limitation has been added to the readme, and a check has been added
to the pipeline snakefile to throw an error when overlapping sample
names are detected.