Remove directory output from fastqc
The directory output for the fastqc tasks is causing issues on the
shared file system of the cluster, since it cannot properly determine
the age of the folder. As a result, it re-runs the fastqc tasks every
time a workflow is restarted, regardless of whether the task has already
completed.

To prevent this, a single dummy output file '.done' has been added to
the fastqc tasks which will be written when fastqc exits successfully.