Use targets file as baitsfile if it is missing
Picard HsMetrics requires both a file containing the targets of the
capture, as well as the actual baits used in the capture. However,
Agilent is refusing to share the baits of their older capture kits.

A common workaround for this issue is to simply specify the targets
file to picard twice.