Rewrite multi sample calling to use scatters
Previously, multi sample calling was done on the merged per sample g.vcf
files. However, in rare cases, VCF files that are merged by bcftools can
trigger a crash in GATK.

To resolve this, the multi sample calling now uses the scattered per
sample g.vcf files, so that GATK does not have to act on VCF files that
were modified by bcftools.

As an added advantage, using the scatters directly for the multi sample
calling allows for better parallelisation.